+++ homework 3/src/app.ts
@@ -8,17 +8,23 @@ import { init } from './init/init';
 import {
   customLoggerMiddleware,
   winstonErrorLoggerMiddleware,
-} from './logger/logger';
+} from './utils/logger';
 import { PORT as APP_PORT } from './configs/config';
 import { checkAccessTokenMiddleware } from './routers/login/tokens';
+import { validatorError } from './utils/validator';
 
+const PORT = (process.env.PORT && parseInt(process.env.PORT)) || APP_PORT;
 const app: Application = express();
 
 app.use(express.json());
 
 app.use(customLoggerMiddleware);
 
-app.use(cors());
+app.use(
+  cors({
+    origin: `http://localhost:${PORT}`,
+  })
+);
 
 app.use('', loginRouter);
 
@@ -28,21 +34,9 @@ app.use('/users', userRouter);
 
 app.use('/groups', groupRouter);
 
-// TODO: Вынести
-app.use((err: any, req: Request, res: Response) => {
-  if (err && err.error && err.error.isJoi) {
-    // return json response with status 400 if there is a joi error,
-    res.status(400).json({
-      type: err.type,
-      message: err.error.toString(),
-    });
-  }
-  res.status(500).send(err);
-});
+app.use(validatorError);
 
 app.use(winstonErrorLoggerMiddleware);
-
-const PORT = (process.env.PORT && parseInt(process.env.PORT)) || APP_PORT;
 
 sequelize
   .authenticate()
